Extend and slightly modify API & add documentation
This is a BREAKING CHANGE.

The `ToHex` trait now requires the `write_hex` method and
doesn't have a `to_hex` method anymore. It makes sense to make
the required methods as general as possible. In this case,
making the method converting to `String` required and
`write_hex` provided doesn't make a lot of sense. Now users can
use the `trait` if they need maximum control.

A `write_upper_hex` method was also added, if users prefer
or need to use uppercase characters.

`FromHex` is now also implemented for small fixed sized
arrays. This results in a new possible error case which was
added to `FromHexError`.

Three handy functions were added to easy usage of this
crate. Now the functions cover most use cases and the traits
can be used in specialized situations.

Lastly, documentation including examples was added to
everything (basically).