feat: `decode_in_slice()`
For memory constrained enviroments and network protocols being able to use
memory without extra copies is a benefit. E.g. in my ZMODEM2 crate I need
to do this for hex messages:

```
if encoding == Encoding::ZHEX {
    let mut tmp = [0; (FrameHeader::unescaped_size(Encoding::ZHEX) - 1) / 2];
    match hex::decode_to_slice(out, &mut tmp) {
        Ok(x) => x,
        _ => {
            log::error!("from_hex error");
            return Err(ErrorKind::InvalidData.into());
        }
    }
    out.clear();
    out.extend_from_slice(&tmp);
}
```

It is good example since the protocol frames can come also in binary format
in addition to hex. It would be nice to do this preprocessing within the
same buffer as where the data landed to make a transparent zerocopy flow.

Link: https://github.com/jarkkojs/zmodem2/
Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>