fix: make FromHexError work with no_std thiserror
`thiserror` requires `Eq` and an `Error` trait in order
to derive an error like this:

```
#[error("Hex error: {0}")]
HexError(#[from] hex::FromHexError),
```

Now that `Error` is part of core, that `Error` trait can
be used instead of the one in `std`. This allows for better
support in the `no_std` use case.

See: https://github.com/rust-lang/rust/issues/103765